[XPU] mv pre/post process
